debian-common: Add fdisk package as a dependency in wic
In downstream when we build the bullseye images for de0-nano-soc target it's
getting failed in wic image creation step due to sfdisk required to build
the wic images.

To reproduce this issue in upstream we triggered the de0-nano-soc build
with --system-id option enabled in wks.in file, this issue present in
upstream build hence added the dependency package fdisk to the wic
imager install.

Ref commit for de0-nano-soc.wks.in: c8eb30bfbff4

For debian-stretch added an exclusion due to no 'fdisk' package in it.

Signed-off-by: Srinuvasan A <srinuvasan_a@mentor.com>
Signed-off-by: Uladzimir Bely <ubely@ilbers.de>