example-raw: Download example-raw source files.
example-raw-0.3 used to generated its files using "echo >" during
'install' task. So, potential issues in 'fetch', 'unpack' and 'patch'
were not checked in Isar

This patch produces the similar example-raw package, but not it
uses fetching from a remote server ant patching of the code in order
to cover all the tasks in queue.

Signed-off-by: Uladzimir Bely <ubely@ilbers.de>