Add ISAR version of biosplusefi WIC plugin
This patch adds support to create a disk image that works with both
EFI and legacy bios. The biosplusefi-isar is based on the original
biosplusefi WIC plugin that internally calls the bootimg-pcbios and
bootimg-efi. By that, code duplication can be avoided.

The generated image can be booted with EFI or legacy pcbios.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>