Add pcbios option to start_vm
This patch adds the --pcbios option to the start_vm script to
enforce that any -bios QEMU options are removed.
By that, QEMU is forced to use a pc bios (like seabios).

This helps to test combined efi + pcbios disks as these can be
bootet with both efi and pc bios.

Signed-off-by: Felix Moessbauer <felix.moessbauer@siemens.com>